[Refactor] #49 리뷰순 평점순 정렬 구조 변경 및 페이지네이션 적용

+++ src/main/java/com/lokoko/domain/product/dto/ProductSummary.java
@@ -1,10 +1,8 @@
 package com.lokoko.domain.product.dto;
 
-import java.math.BigDecimal;
-
 public record ProductSummary(
         String imageUrl,
         Long reviewCount,
-        BigDecimal avgRating
+        Double avgRating
 ) {
 }
